scripts/requirements.txt
google-genai
pymupdf
pillow
python-dotenv